#557 Viewer Panel Pdfs - support with draw layers
