+++ src/essence/Basics/Map_/Map_.js
@@ -698,19 +698,13 @@ function featureDefaultClick(feature, layer, e) {
     }
 
     function keepGoing() {
-        //View images
-        var propImages = propertiesToImages(
-            feature.properties,
-            layer.options.metadata ? layer.options.metadata.base_url || '' : ''
-        )
-
         Kinds.use(
             L_.layers.data[layer.options.layerName].kind,
             Map_,
             feature,
             layer,
             layer.options.layerName,
-            propImages,
+            null,
             e
         )
 
@@ -724,13 +718,7 @@ function featureDefaultClick(feature, layer, e) {
             }
         }
 
-        Viewer_.changeImages(propImages, feature, layer)
-        for (var i in propImages) {
-            if (propImages[i].type == 'radargram') {
-                //Globe_.radargram( layer.options.layerName, feature.geometry, propImages[i].url, propImages[i].length, propImages[i].depth );
-                break
-            }
-        }
+        Viewer_.changeImages(feature, layer)
 
         //figure out how to construct searchStr in URL. For example: a ChemCam target can sometime
         //be searched by "target sol", or it can be searched by "sol target" depending on config file.
@@ -1189,98 +1177,6 @@ function allLayersLoaded() {
             }
         }
     }
-}
-
-function propertiesToImages(props, baseUrl) {
-    baseUrl = baseUrl || ''
-    var images = []
-    //Use "images" key first
-    if (props.hasOwnProperty('images')) {
-        for (var i = 0; i < props.images.length; i++) {
-            if (props.images[i].url) {
-                var url = baseUrl + props.images[i].url
-                if (!F_.isUrlAbsolute(url)) url = L_.missionPath + url
-                if (props.images[i].isModel) {
-                    images.push({
-                        url: url,
-                        texture: props.images[i].texture,
-                        name:
-                            (props.images[i].name ||
-                                props.images[i].url.match(/([^\/]*)\/*$/)[1]) +
-                            ' [Model]',
-                        type: 'model',
-                        isPanoramic: false,
-                        isModel: true,
-                        values: props.images[i].values || {},
-                        master: props.images[i].master,
-                    })
-                } else {
-                    if (props.images[i].isPanoramic) {
-                        images.push({
-                            ...props.images[i],
-                            url: url,
-                            name:
-                                (props.images[i].name ||
-                                    props.images[i].url.match(
-                                        /([^\/]*)\/*$/
-                                    )[1]) + ' [Panoramic]',
-                            type: 'photosphere',
-                            isPanoramic: true,
-                            isModel: false,
-                            values: props.images[i].values || {},
-                            master: props.images[i].master,
-                        })
-                    }
-                    images.push({
-                        url: url,
-                        name:
-                            props.images[i].name ||
-                            props.images[i].url.match(/([^\/]*)\/*$/)[1],
-                        type: props.images[i].type || 'image',
-                        isPanoramic: false,
-                        isModel: false,
-                        values: props.images[i].values || {},
-                        master: props.images[i].master,
-                    })
-                }
-            }
-        }
-    }
-    //If there isn't one, search all string valued props for image urls
-    else {
-        for (var p in props) {
-            if (
-                typeof props[p] === 'string' &&
-                props[p].toLowerCase().match(/\.(jpeg|jpg|gif|png|xml)$/) !=
-                    null
-            ) {
-                var url = props[p]
-                if (!F_.isUrlAbsolute(url)) url = L_.missionPath + url
-                images.push({
-                    url: url,
-                    name: p,
-                    isPanoramic: false,
-                    isModel: false,
-                })
-            }
-            if (
-                typeof props[p] === 'string' &&
-                (props[p].toLowerCase().match(/\.(obj)$/) != null ||
-                    props[p].toLowerCase().match(/\.(dae)$/) != null)
-            ) {
-                var url = props[p]
-                if (!F_.isUrlAbsolute(url)) url = L_.missionPath + url
-                images.push({
-                    url: url,
-                    name: p,
-                    isPanoramic: false,
-                    isModel: true,
-                })
-            }
-        }
-    }
-
-    return images
 }
 
 function buildToolBar() {
